Styling maciek

+++ src/app/pages/page/page.component.ts
@@ -1,12 +1,19 @@
-import { Component, OnInit } from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
+import {Component, ElementRef, OnInit} from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
 import {PageService} from "../../services/page.service";
 import {Location} from "@angular/common";
 import {MenuService} from "../../services/menu.service";
 import {PageDTO} from "../../services/menu.service";
 import {MatDialog} from "@angular/material/dialog";
 import {TagsComponent} from "../../tags/tags.component";
-
+import {
+  MatSnackBar,
+  MatSnackBarHorizontalPosition,
+  MatSnackBarVerticalPosition,
+} from '@angular/material/snack-bar';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {delay} from "rxjs";
+import {DeleteDialogComponent, DeleteDialogData} from "../../delete-dialog/delete-dialog.component";
 
 interface PageContent {
   latex: boolean,
@@ -29,17 +36,24 @@ export class PageComponent implements OnInit {
   editPageField:string;
   deletePageField:string;
   editTagsField:string;
+  horizontalPosition: MatSnackBarHorizontalPosition = 'end';
+  verticalPosition: MatSnackBarVerticalPosition = 'bottom';
+  pageName:string;
+  editMode: boolean;
 
   constructor(
     private route: ActivatedRoute,
     private pageService: PageService,
     private location: Location,
     private menuService: MenuService,
-    private matDialog: MatDialog
+    private matDialog: MatDialog,
+    private _snackbar:MatSnackBar,
+    private router: Router
   ) {
   }
 
   ngOnInit(): void {
+    this.editMode = false;
     this.id = this.route.snapshot.params['id'];
     this.language = this.route.snapshot.params['lang'];
     this.route.params.subscribe(val => {
@@ -47,7 +61,7 @@ export class PageComponent implements OnInit {
       this.pageContent = [];
       this.getPage();
     });
-    this.setLanguageFields()
+    this.setLanguageFields();
   }
 
   setLanguageFields() {
@@ -69,12 +83,23 @@ export class PageComponent implements OnInit {
         break;
     }
   }
-
+  changeEditMode(newPageName: string){
+    this.editMode = !this.editMode;
+    if(newPageName !== this.pageName){
+      this.pageName = newPageName;
+      const newPage:PageDTO = {
+        pageName: this.pageName,
+        id:this.id
+      }
+      this.pageService.patchPage(newPage).subscribe();
+    }
+  }
   getPage(): void {
     this.pageService.getPage(this.id).subscribe( {
       next: (page: PageDTO) => {
           this.pageHTML = page.content || "";
           this.tags = page.tags;
+          this.pageName = page.pageName || "";
           },
       error: () => {},
       complete: () => {
@@ -86,23 +111,63 @@ export class PageComponent implements OnInit {
     });
   }
 
+  deletePage(evt: MouseEvent){
+    const elementTargeted = new ElementRef(evt.currentTarget);
+    let dialogData:DeleteDialogData = {
+      isPage: true,
+      name: this.pageName,
+      confirmDeleted:false,
+      target: elementTargeted,
+      isLeft:false
+    }
+    let deleteDialogRef = this.matDialog.open(DeleteDialogComponent, {
+      minHeight: '110px',
+      minWidth: '200px',
+      panelClass: 'custom-dialog-container',
+      data: dialogData,
+    });
+    deleteDialogRef.afterClosed().subscribe(
+      result => {
+        if(result){
+          this.pageService.deletePage(this.id).subscribe(
+            () => this.router.navigateByUrl(`/${this.language}/home`)
+          )
+        }
+      }
+    )
+  }
 
-  deletePage(){
-    this.pageService.deletePage(this.id).subscribe(
-      () => {this.menuService.getMenuData();
-        window.location.reload();
-      },
-      )
+  openSnackbar(message:string){
+    this._snackbar.open(message,'', {
+      horizontalPosition: this.horizontalPosition,
+      verticalPosition: this.verticalPosition,
+      duration: 3000,
+      panelClass:['snackbarStyles'],
+    })
   }
 
   editTags(){
     const dialogRef = this.matDialog.open(TagsComponent, {
-      width: "30vw",
-      height:"60vh",
+      width: "400px",
+      minHeight:"300px",
+      maxHeight:"700px",
       data: { tags: this.tags, id:this.id},
     });
-    dialogRef.afterClosed().subscribe(result =>
-      window.location.reload()
+    dialogRef.afterClosed().subscribe(
+      data => {
+        if(data.confirmation){
+        const newPage:PageDTO = {
+          tags:data.tags,
+          id:this.id,
+        }
+        this.pageService.patchPage(newPage).subscribe(
+          () => {
+            this.getPage();
+            this.openSnackbar("You edited Tags");
+          }
+        )
+      }
+      }
     )
   }
 }
